Refactor namings for payment and notification services
