+++ services/config-server/src/main/resources/configurations/payment-service.yml
@@ -15,7 +15,7 @@ spring:
       key-serializer: org.apache.kafka.common.serialization.StringSerializer
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
       properties:
-        spring.json.type.mapping: paymentConfirmation:dev.berke.app.kafka.PaymentNotificationRequest
+        spring.json.type.mapping: paymentConfirmation:dev.berke.app.kafka.PaymentConfirmRequest
 iyzipay:
   api:
     key: sandbox-q6L1GCFJrBE3OUTW8lxbPS5bCDzqeBZZ
