+++ services/notification/src/main/java/dev/berke/app/kafka/payment/PaymentConfirmRequest.java
@@ -1,10 +1,8 @@
-package dev.berke.app.kafka;
-
-import dev.berke.app.payment.PaymentMethod;
+package dev.berke.app.kafka.payment;
 
 import java.math.BigDecimal;
 
-public record PaymentNotificationRequest(
+public record PaymentConfirmRequest(
         String customerName,
         String email,
         BigDecimal totalPrice,
