+++ services/payment/src/main/java/dev/berke/app/basket/BasketClient.java
@@ -4,10 +4,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-@FeignClient(
-        name = "basket-service",
-        url = "${application.config.basket-url}"
-)
+@FeignClient(name = "basket-service", url = "${application.config.basket-url}")
 public interface BasketClient {
 
     @GetMapping("/{customer-id}")
