+++ services/payment/src/main/java/dev/berke/app/kafka/PaymentConfirmRequest.java
@@ -1,8 +1,10 @@
-package dev.berke.app.kafka.payment;
+package dev.berke.app.kafka;
+
+import dev.berke.app.payment.PaymentMethod;
 
 import java.math.BigDecimal;
 
-public record PaymentConfirmation(
+public record PaymentConfirmRequest(
         String customerName,
         String email,
         BigDecimal totalPrice,
