+++ services/payment/src/main/java/dev/berke/app/kafka/PaymentProducer.java
@@ -11,19 +11,19 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class PaymentNotificationProducer {
+public class PaymentProducer {
 
     // KafkaTemplate for sending messages to Kafka broker
-    private final KafkaTemplate<String, PaymentNotificationRequest> kafkaTemplate;
+    private final KafkaTemplate<String, PaymentConfirmRequest> kafkaTemplate;
 
     // Send a payment notification to the Kafka topic
-    public void sendPaymentNotification(PaymentNotificationRequest paymentNotificationRequest) {
-        log.info("Sending notification with body <{}>", paymentNotificationRequest);
+    public void sendPaymentNotification(PaymentConfirmRequest paymentConfirmRequest) {
+        log.info("Sending notification with body <{}>", paymentConfirmRequest);
 
         // Create a message to be sent to Kafka
-        Message<PaymentNotificationRequest> message = MessageBuilder
+        Message<PaymentConfirmRequest> message = MessageBuilder
                 // Sets the payload of the message to the provided request
-                .withPayload(paymentNotificationRequest)
+                .withPayload(paymentConfirmRequest)
                 .setHeader(KafkaHeaders.TOPIC, "payment-topic")
                 .build();
